backend/requirements.txt
# Wired-Part Backend Dependencies
# =================================
# Uses >= instead of == for Python 3.14 compatibility

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
aiosqlite>=0.20.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0

# Utilities
python-multipart>=0.0.20
